+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - NEW `--prettier` argument to scaffold a new app with prettier and installs a precommit hook
 - NEW `--ssl` argument to scaffold a new app with https express server
 - DEPRECATED main.ts, use @angular/cli supplied src/main.ts instead
+- FIXED issue that prevented library build from compiling global sass
 
 NOTE: Running the https express server requires a public key and certificate. Generate `key.pem` and `cert.pem` and save in the `backend` directory.